// utils/state/action-types.ts
export enum ActionTypes {
  SET_STATE = 'SET_STATE',
  CLEAR_STATE = 'CLEAR_STATE',
  SET_ONBOARDING_COMPLETED = 'SET_ONBOARDING_COMPLETED',
  SET_FIRST_NAME = 'SET_FIRST_NAME',
  SET_EMAIL = 'SET_EMAIL',
  SET_LAST_NAME = 'SET_LAST_NAME',
  SET_AVATAR = 'SET_AVATAR',
  SET_PHONE = 'SET_PHONE',
  SET_NOTIFICATIONS = 'SET_NOTIFICATIONS',
  SET_MENU_LIST = 'SET_MENU_LIST',
  SET_CATEGORY_FILTER = 'SET_CATEGORY_FILTER',
}